moved stuff to blockchain state struct
